[iOS] chore: Settingview 분리
